refactor: :art: moves Row components into the Table directory
