--- frontend/src/components/Agreements/AgreementTableRow/index.js
@@ -1,2 +1,0 @@
-export * from "./AgreementTableRow";
-export { default } from "./AgreementTableRow";
