--- frontend/src/components/BudgetLineItems/AllBLIRow/index.js
@@ -1,1 +1,0 @@
-export { default } from "./AllBLIRow";
