+++ frontend/src/components/BudgetLineItems/AllBudgetLinesTable/AllBudgetLinesTable.js
@@ -3,7 +3,7 @@ import { useNavigate } from "react-router-dom";
 import PropTypes from "prop-types";
 import _ from "lodash";
 import Table from "../../UI/Table";
-import AllBLIRow from "../AllBLIRow";
+import AllBLIRow from "./AllBLIRow";
 import { All_BUDGET_LINES_TABLE_HEADINGS, BLIS_PER_PAGE } from "../../../constants";
 import PaginationNav from "../../UI/PaginationNav/PaginationNav";
