--- frontend/src/components/BudgetLineItems/BLIRow/index.js
@@ -1,1 +1,0 @@
-export { default } from "./BLIRow";
