+++ frontend/src/components/BudgetLineItems/BudgetLinesTable/BudgetLinesTable.jsx
@@ -3,7 +3,7 @@ import TotalSummaryCard from "../TotalSummaryCard/TotalSummaryCard";
 import Table from "../../UI/Table";
 import "./BudgetLinesTable.scss";
 import { BUDGET_LINE_TABLE_HEADERS } from "../../../constants";
-import BLIRow from "../BLIRow";
+import BLIRow from "./BLIRow";
 
 /**
  * A table component that displays budget lines.
